DEVTOOLS: COMPANION: Finish integrating create_classicmacfonts as createmacfonts
